pyproject.toml
[tool.poetry]
name = "monopoly4u-game"
version = "0.1.0"
description = ""
authors = ["juhc <chalkov2002@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-socketio = "^5.11.2"
setuptools = "^70.0.0"
redis = "^5.0.4"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
